vs2022 c++20(143) wxWidgets Version: 3.2.6

+++ src/GRPluginManager.cpp
@@ -38,7 +38,7 @@ GRPluginManager::GRPluginManager()
 /*------------------------------------------------------------------------------*/
 GRPluginManager::~GRPluginManager()
 {
-	for(unsigned int x = 0; x < Plugins.size(); x++) {
+	for(size_t x = 0; x < Plugins.size(); x++) {
 		delete(Plugins[x]);
 	}
 	Plugins.clear();
